+++ Controllers/Bigkala1/Bigkala1Controller.js.js
@@ -34,7 +34,7 @@ exports.getCalculatedData = (req,res) => {
     const scooterTotal = totalAmountOfWeek > 0 ? (((totalAmountOfWeek * 0.50) / 2) + (Math.abs(lastWeekMakeUp) / 2)).toFixed(2) : (totalEarned / 2)
     const bigkala1Total = totalAmountOfWeek > 0 ? (totalAmountOfWeek * 0.50).toFixed(2) : 0
     const makeUp = totalAmountOfWeek > 0 ? 0 : totalAmountOfWeek
-    const headFees = (playerCount.find((x) => x.agentName === "BIGKALA1") && headfees.hasOwnProperty("BIGKALA1")) ? parseInt(headfees["BIGKALA1"] * playerCount.find((x) => x.agentName === "BIGKALA1")["playersCount"]).toFixed(2) : 0
+    const headFees = (playerCount.find((x) => x.agentName === "BIGKALA1") && headfees.hasOwnProperty("BIGKALA1")) ? parseFloat(headfees["BIGKALA1"] * playerCount.find((x) => x.agentName === "BIGKALA1")["playersCount"]).toFixed(2) : 0
     const scooterNet = scooterTotal - headFees
 
     const data = {
