+++ Controllers/Dso/DsoController.js
@@ -28,7 +28,7 @@ exports.getCalculatedData = (req,res) => {
     const totalAmountOfWeek = result[Object.keys(result)[0]].filter((x, i) => i !== 0).map((y) => y["F"]).reduce((a, b) => a + b, 0)
     const scooterTotal = totalAmountOfWeek / 2
     const dsoTotal = totalAmountOfWeek / 2
-    const headFees = (playerCount.find((x) => x.agentName === "DSO") && headfees.hasOwnProperty("DSO")) ? parseInt(headfees["DSO"] * playerCount.find((x) => x.agentName === "DSO")["playersCount"]).toFixed(2) : 0
+    const headFees = (playerCount.find((x) => x.agentName === "DSO") && headfees.hasOwnProperty("DSO")) ? parseFloat(headfees["DSO"] * playerCount.find((x) => x.agentName === "DSO")["playersCount"]).toFixed(2) : 0
     const scooterNet = scooterTotal - headFees
 
     const data = {
