+++ Controllers/Jd30/Jd30Controller.js
@@ -34,7 +34,7 @@ exports.getCalculatedData = (req,res) => {
     const scooterTotal = totalAmountOfWeek > 0 ? ((totalAmountOfWeek * 0.50) + Math.abs(lastWeekMakeUp)).toFixed(2) : totalEarned
     const jd30Total = totalAmountOfWeek > 0 ? (totalAmountOfWeek * 0.50).toFixed(2) : 0
     const makeUp = totalAmountOfWeek > 0 ? 0 : totalAmountOfWeek
-    const headFees = (playerCount.find((x) => x.agentName === "JD30") && headfees.hasOwnProperty("JD30")) ? parseInt(headfees["JD30"] * playerCount.find((x) => x.agentName === "JD30")["playersCount"]).toFixed(2) : 0
+    const headFees = (playerCount.find((x) => x.agentName === "JD30") && headfees.hasOwnProperty("JD30")) ? parseFloat(headfees["JD30"] * playerCount.find((x) => x.agentName === "JD30")["playersCount"]).toFixed(2) : 0
     const scooterNet = scooterTotal - headFees
 
     const data = {
