+++ Controllers/Joeyspotsa/JoeyspotsaController.js
@@ -28,7 +28,7 @@ exports.getCalculatedData = (req,res) => {
     const totalAmountOfWeek = result[Object.keys(result)[0]].filter((x, i) => i !== 0).map((y) => y["F"]).reduce((a, b) => a + b, 0)
     const scooterTotal = totalAmountOfWeek > 0 ? (totalAmountOfWeek * 0.75).toFixed(2) : totalAmountOfWeek
     const joeyspotsaTotal = totalAmountOfWeek > 0 ? (totalAmountOfWeek * 0.25).toFixed(2) : 0
-    const headFees = (playerCount.find((x) => x.agentName === "JOEYSPOTSA") && headfees.hasOwnProperty("JOEYSPOTSA")) ? parseInt(headfees["JOEYSPOTSA"] * playerCount.find((x) => x.agentName === "JOEYSPOTSA")["playersCount"]).toFixed(2) : 0
+    const headFees = (playerCount.find((x) => x.agentName === "JOEYSPOTSA") && headfees.hasOwnProperty("JOEYSPOTSA")) ? parseFloat(headfees["JOEYSPOTSA"] * playerCount.find((x) => x.agentName === "JOEYSPOTSA")["playersCount"]).toFixed(2) : 0
     const scooterNet = scooterTotal - headFees
 
     const data = {
