+++ Controllers/M508/M508Controller.js
@@ -28,7 +28,7 @@ exports.getCalculatedData = (req,res) => {
     const totalAmountOfWeek = result[Object.keys(result)[0]].filter((x, i) => i !== 0).map((y) => y["F"]).reduce((a, b) => a + b, 0)
     const scooterTotal = totalAmountOfWeek > 0 ? ((totalAmountOfWeek * 0.75) / 2).toFixed(2) : (totalAmountOfWeek / 2)
     const m508Total = totalAmountOfWeek > 0 ? (totalAmountOfWeek * 0.25).toFixed(2) : 0
-    const headFees = (playerCount.find((x) => x.agentName === "M508") && headfees.hasOwnProperty("M508")) ? parseInt(headfees["M508"] * playerCount.find((x) => x.agentName === "M508")["playersCount"]).toFixed(2) : 0
+    const headFees = (playerCount.find((x) => x.agentName === "M508") && headfees.hasOwnProperty("M508")) ? parseFloat(headfees["M508"] * playerCount.find((x) => x.agentName === "M508")["playersCount"]).toFixed(2) : 0
     const scooterNet = scooterTotal - headFees
 
     const data = {
