+++ Controllers/Sc222/Sc222Controller.js
@@ -34,7 +34,7 @@ exports.getCalculatedData = (req,res) => {
   const scooterTotal = totalAmountOfWeek > 0 ? ((totalAmountOfWeek * 0.70) + Math.abs(lastWeekMakeUp)).toFixed(2) : totalEarned
   const sc222Total = totalAmountOfWeek > 0 ? (totalAmountOfWeek * 0.30).toFixed(2) : 0
   const makeUp = totalAmountOfWeek > 0 ? 0 : totalAmountOfWeek
-  const headFees = (playerCount.find((x) => x.agentName === "SCC222") && headfees.hasOwnProperty("SCC222")) ? parseInt(headfees["SCC222"] * playerCount.find((x) => x.agentName === "SCC222")["playersCount"]).toFixed(2) : 0
+  const headFees = (playerCount.find((x) => x.agentName === "SCC222") && headfees.hasOwnProperty("SCC222")) ? parseFloat(headfees["SCC222"] * playerCount.find((x) => x.agentName === "SCC222")["playersCount"]).toFixed(2) : 0
   const scooterNet = scooterTotal - headFees
 
   const data = {
