+++ Controllers/Zebrafish/ZebrafishController.js
@@ -34,7 +34,7 @@ exports.getCalculatedData = (req,res) => {
   const scooterTotal = totalAmountOfWeek > 0 ? ((totalAmountOfWeek * 0.70) + Math.abs(lastWeekMakeUp)).toFixed(2) : totalEarned
   const zebrafishTotal = totalAmountOfWeek > 0 ? (totalAmountOfWeek * 0.30).toFixed(2) : 0
   const makeUp = totalAmountOfWeek > 0 ? 0 : totalAmountOfWeek
-  const headFees = (playerCount.find((x) => x.agentName === "ZEBRAFISH") && headfees.hasOwnProperty("ZEBRAFISH")) ? parseInt(headfees["ZEBRAFISH"] * playerCount.find((x) => x.agentName === "ZEBRAFISH")["playersCount"]).toFixed(2) : 0
+  const headFees = (playerCount.find((x) => x.agentName === "ZEBRAFISH") && headfees.hasOwnProperty("ZEBRAFISH")) ? parseFloat(headfees["ZEBRAFISH"] * playerCount.find((x) => x.agentName === "ZEBRAFISH")["playersCount"]).toFixed(2) : 0
   const scooterNet = scooterTotal - headFees
 
   const data = {
